requirements.txt
# General Modules
django==3.2.10
python-decouple
pillow
djangorestframework
coreapi
djangorestframework-simplejwt

# Third party modules
drf-yasg[validation]
markdown
django-filter

# Deployment modules


# Email Third party modules
django-mail-templated


# formatting and testing modules
black
flake8